refactor: update of component name
Task 18
