+++ package.json
@@ -86,7 +86,16 @@
     "ts-jest": "^29.0.5",
     "typescript": "^4.5.5",
     "vite": "^4.0.4"
+  },
+  "dependencies": {
+    "@ark-ui/react": "^0.7.2",
+    "@chakra-ui/react": "^2.7.1",
+    "@emotion/styled": "^11.11.0",
+    "@tanstack/react-table": "^8.9.3",
+    "chakra-react-select": "^4.6.0",
+    "framer-motion": "^10.12.18",
+    "i18next": "^23.2.8",
+    "react-i18next": "^13.0.1",
+    "react-icons": "^4.10.1"
   }
 }
-
-
