+++ src/App.tsx
@@ -2,12 +2,12 @@ import React, { Box } from '@chakra-ui/react';
 
 import DataTable from './components/DataTable';
 import ThemeToggleButton from './components/globalpartials/ThemeToggleButton';
-import BodyWrapper from './components/BodyWrapper';
-import FunctionalWrapper from './components/FunctionalWrapper';
+import Table from './components/Table';
+import Header from './components/Header';
 import TableTitle from './components/functionalcomponents/TableTitle';
 import GlobalSearch from './components/functionalcomponents/GlobalSearch';
-import FooterWrapper from './components/FooterWrapper';
-import PaginationWrapper from './components/footercomponents/PaginationWrapper';
+import Footer from './components/Footer';
+import Pagination from './components/footercomponents/Pagination';
 import SelectedNumber from './components/footercomponents/SelectedNumber';
 import columns from './components/ProvideByConsumer/Columns';
 import PageSizeControl from './components/footercomponents/PageSizeControl';
@@ -18,20 +18,20 @@ function App(): JSX.Element {
   return (
     <Box>
       <DataTable>
-        <FunctionalWrapper>
+        <Header>
           <TableTitle>Member</TableTitle>
           <GlobalSearch />
-        </FunctionalWrapper>
-        <BodyWrapper columns={columns}>
+        </Header>
+        <Table columns={columns}>
           <TableHeader />
           <TableBody />
-        </BodyWrapper>
-        <FooterWrapper>
-          <PaginationWrapper>
+        </Table>
+        <Footer>
+          <Pagination>
             <SelectedNumber />
             <PageSizeControl pages={[5, 10, 25, 50, 100]} />
-          </PaginationWrapper>
-        </FooterWrapper>
+          </Pagination>
+        </Footer>
       </DataTable>
       <ThemeToggleButton pos="fixed" bottom="2" right="2" />
     </Box>
