+++ src/export.ts
@@ -2,18 +2,18 @@
 export { default as DataTable } from './components/DataTable';
 
 //export FunctionalComponents
-export { default as FunctionalWrapper } from './components/FunctionalWrapper';  // FunctionalWapper
+export { default as Header } from './components/Header'; // Header
 export { default as GlobalSearch } from './components/functionalcomponents/GlobalSearch'; // Searchbar component
 export { default as TableTitle } from './components/functionalcomponents/TableTitle'; // TitleComponent
 
 //export BodyComponents
-export { default as BodyWrapper } from './components/BodyWrapper'; // BodyWrapper component
+export { default as Table } from './components/Table'; // Table component
 export { default as TableHeader } from './components/bodycomponents/TableHeader'; // TableHeader Component
-export { default as TableBody } from './components/bodycomponents/TableBody';  // TableBody Component
+export { default as TableBody } from './components/bodycomponents/TableBody'; // TableBody Component
 
 //export Footer components
-export { default as FooterWrapper } from './components/FooterWrapper'; // FooterWapper component
-export { default as PaginationWapper } from './components/footercomponents/PaginationWrapper'; // PaginationWapper Component 
+export { default as Footer } from './components/Footer'; // Footer component
+export { default as Pagination } from './components/footercomponents/Pagination'; // Pagination Component
 export { default as PageSizeControl } from './components/footercomponents/PageSizeControl'; // PageSizeControl Component
 export { default as SelectedNumber } from './components/footercomponents/SelectedNumber'; // SelectedNumber Component
