+++ src/stories/DataTable.stories.tsx
@@ -3,12 +3,12 @@ import type { Meta } from '@storybook/react';
 
 import DefaultDataTable from './DefaultDataTable';
 import DataTable from '../components/DataTable';
-import BodyWrapper from '../components/BodyWrapper';
-import FunctionalWrapper from '../components/FunctionalWrapper';
+import Table from '../components/Table';
+import Header from '../components/Header';
 import TableTitle from '../components/functionalcomponents/TableTitle';
 import GlobalSearch from '../components/functionalcomponents/GlobalSearch';
-import FooterWrapper from '../components/FooterWrapper';
-import PaginationWrapper from '../components/footercomponents/PaginationWrapper';
+import Footer from '../components/Footer';
+import Pagination from '../components/footercomponents/Pagination';
 import SelectedNumber from '../components/footercomponents/SelectedNumber';
 import columns from '../components/ProvideByConsumer/Columns';
 import PageSizeControl from '../components/footercomponents/PageSizeControl';
@@ -40,20 +40,20 @@ export const DefaultTable = () => (
   <ChakraProvider theme={theme}>
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <DataTable>
-      <FunctionalWrapper>
+      <Header>
         <TableTitle>Default DataTable</TableTitle>
         <GlobalSearch />
-      </FunctionalWrapper>
-      <BodyWrapper columns={columns}>
+      </Header>
+      <Table columns={columns}>
         <TableHeader />
         <TableBody />
-      </BodyWrapper>
-      <FooterWrapper>
-        <PaginationWrapper>
+      </Table>
+      <Footer>
+        <Pagination>
           <SelectedNumber />
           <PageSizeControl pages={[5, 10, 25, 50, 100]} />
-        </PaginationWrapper>
-      </FooterWrapper>
+        </Pagination>
+      </Footer>
     </DataTable>
   </ChakraProvider>
 );
@@ -62,19 +62,19 @@ export const WithoutGlobalSearch = () => (
   <ChakraProvider theme={theme}>
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <DataTable>
-      <FunctionalWrapper>
+      <Header>
         <TableTitle>DataTable without Global search</TableTitle>
-      </FunctionalWrapper>
-      <BodyWrapper columns={columns}>
+      </Header>
+      <Table columns={columns}>
         <TableHeader />
         <TableBody />
-      </BodyWrapper>
-      <FooterWrapper>
-        <PaginationWrapper>
+      </Table>
+      <Footer>
+        <Pagination>
           <SelectedNumber />
           <PageSizeControl pages={[5, 10, 25, 50, 100]} />
-        </PaginationWrapper>
-      </FooterWrapper>
+        </Pagination>
+      </Footer>
     </DataTable>
   </ChakraProvider>
 );
@@ -83,18 +83,18 @@ export const WithoutPageSelection = () => (
   <ChakraProvider theme={theme}>
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <DataTable>
-      <FunctionalWrapper>
+      <Header>
         <TableTitle>DataTable without page selection</TableTitle>
-      </FunctionalWrapper>
-      <BodyWrapper columns={columns}>
+      </Header>
+      <Table columns={columns}>
         <TableHeader />
         <TableBody />
-      </BodyWrapper>
-      <FooterWrapper>
-        <PaginationWrapper>
+      </Table>
+      <Footer>
+        <Pagination>
           <SelectedNumber />
-        </PaginationWrapper>
-      </FooterWrapper>
+        </Pagination>
+      </Footer>
     </DataTable>
   </ChakraProvider>
 );
@@ -103,19 +103,19 @@ export const WithoutTableHeader = () => (
   <ChakraProvider theme={theme}>
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <DataTable>
-      <FunctionalWrapper>
+      <Header>
         <TableTitle>DataTable without tableheader</TableTitle>
         <GlobalSearch />
-      </FunctionalWrapper>
-      <BodyWrapper columns={columns}>
+      </Header>
+      <Table columns={columns}>
         <TableBody />
-      </BodyWrapper>
-      <FooterWrapper>
-        <PaginationWrapper>
+      </Table>
+      <Footer>
+        <Pagination>
           <SelectedNumber />
           <PageSizeControl pages={[5, 10, 25, 50, 100]} />
-        </PaginationWrapper>
-      </FooterWrapper>
+        </Pagination>
+      </Footer>
     </DataTable>
   </ChakraProvider>
 );
@@ -124,14 +124,14 @@ export const WithoutFooter = () => (
   <ChakraProvider theme={theme}>
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <DataTable>
-      <FunctionalWrapper>
+      <Header>
         <TableTitle>DataTable without footer</TableTitle>
         <GlobalSearch />
-      </FunctionalWrapper>
-      <BodyWrapper columns={columns}>
+      </Header>
+      <Table columns={columns}>
         <TableHeader />
         <TableBody />
-      </BodyWrapper>
+      </Table>
     </DataTable>
   </ChakraProvider>
 );
