+++ src/stories/DataTableWithoutPageSelection.tsx
@@ -1,11 +1,11 @@
 import React, { Box, ChakraProvider, ColorModeScript } from '@chakra-ui/react';
 
 import DataTable from '../components/DataTable';
-import BodyWrapper from '../components/BodyWrapper';
-import FunctionalWrapper from '../components/FunctionalWrapper';
+import Table from '../components/Table';
+import Header from '../components/Header';
 import TableTitle from '../components/functionalcomponents/TableTitle';
-import FooterWrapper from '../components/FooterWrapper';
-import PaginationWrapper from '../components/footercomponents/PaginationWrapper';
+import Footer from '../components/Footer';
+import Pagination from '../components/footercomponents/Pagination';
 import SelectedNumber from '../components/footercomponents/SelectedNumber';
 import columns from '../components/ProvideByConsumer/Columns';
 import theme from '../theme';
@@ -18,18 +18,18 @@ const DataTableWithoutPageSelection = () => {
       <ChakraProvider theme={theme}>
         <ColorModeScript initialColorMode={theme.config.initialColorMode} />
         <DataTable>
-          <FunctionalWrapper>
+          <Header>
             <TableTitle>DataTable without page selection</TableTitle>
-          </FunctionalWrapper>
-          <BodyWrapper columns={columns}>
+          </Header>
+          <Table columns={columns}>
             <TableHeader />
             <TableBody />
-          </BodyWrapper>
-          <FooterWrapper>
-            <PaginationWrapper>
+          </Table>
+          <Footer>
+            <Pagination>
               <SelectedNumber />
-            </PaginationWrapper>
-          </FooterWrapper>
+            </Pagination>
+          </Footer>
         </DataTable>
       </ChakraProvider>
     </Box>
