+++ src/stories/DataTableWithoutTableHeader.tsx
@@ -1,12 +1,12 @@
 import React, { Box, ChakraProvider, ColorModeScript } from '@chakra-ui/react';
 
 import DataTable from '../components/DataTable';
-import BodyWrapper from '../components/BodyWrapper';
-import FunctionalWrapper from '../components/FunctionalWrapper';
+import Table from '../components/Table';
+import Header from '../components/Header';
 import TableTitle from '../components/functionalcomponents/TableTitle';
 import GlobalSearch from '../components/functionalcomponents/GlobalSearch';
-import FooterWrapper from '../components/FooterWrapper';
-import PaginationWrapper from '../components/footercomponents/PaginationWrapper';
+import Footer from '../components/Footer';
+import Pagination from '../components/footercomponents/Pagination';
 import SelectedNumber from '../components/footercomponents/SelectedNumber';
 import columns from '../components/ProvideByConsumer/Columns';
 import PageSizeControl from '../components/footercomponents/PageSizeControl';
@@ -19,19 +19,19 @@ const DataTableWithoutTableHeader = () => {
       <ChakraProvider theme={theme}>
         <ColorModeScript initialColorMode={theme.config.initialColorMode} />
         <DataTable>
-          <FunctionalWrapper>
+          <Header>
             <TableTitle>DataTable without tableheader</TableTitle>
             <GlobalSearch />
-          </FunctionalWrapper>
-          <BodyWrapper columns={columns}>
+          </Header>
+          <Table columns={columns}>
             <TableBody />
-          </BodyWrapper>
-          <FooterWrapper>
-            <PaginationWrapper>
+          </Table>
+          <Footer>
+            <Pagination>
               <SelectedNumber />
               <PageSizeControl pages={[5, 10, 25, 50, 100]} />
-            </PaginationWrapper>
-          </FooterWrapper>
+            </Pagination>
+          </Footer>
         </DataTable>
       </ChakraProvider>
     </Box>
